pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biblio"
version = "0.1.0"
description = "Portable bibliography workspace tool"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Arash Shahidi"}]
dependencies = [
  "pyyaml",
]

[project.optional-dependencies]
bibtex = ["pybtex"]
openalex = ["pybtex", "httpx", "tenacity", "tqdm"]
dev = ["pytest", "pybtex", "httpx", "tenacity", "tqdm"]

[project.scripts]
biblio = "biblio.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["biblio", "biblio.*"]

[tool.setuptools.package-data]
biblio = ["templates/**/*.tmpl"]
